requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
SQLAlchemy==2.0.15
typing_extensions==4.6.2
Werkzeug==2.3.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0